Changed type of numarenas from uint to size_t to silence a GCC warning on 64bit OSes. Reviewed by Benjamin Peterson.

+++ Objects/obmalloc.c
@@ -517,7 +517,7 @@ new_arena(void)
 #endif
 	if (unused_arena_objects == NULL) {
 		uint i;
-		uint numarenas;
+		size_t numarenas;
 		size_t nbytes;
 
 		/* Double the number of arena objects on each allocation.
